requirements.txt
aiohttp
aiosqlite
discord.py
python-dotenv
youtube_search
pynacl